+++ core/core.php
@@ -677,6 +677,7 @@ if ($_POST['action'] === 'update_user_settings') {
     $chem_vs_brand = isset($_POST["chem_vs_brand"]) && $_POST["chem_vs_brand"] === 'true' ? '1' : '0';
     $chkVersion = isset($_POST["chkVersion"]) && $_POST["chkVersion"] === 'true' ? '1' : '0';
     $multi_dim_perc = isset($_POST["multi_dim_perc"]) && $_POST["multi_dim_perc"] === 'true' ? '1' : '0';
+    $allow_incomplete_ingredients = isset($_POST["allow_incomplete_ingredients"]) && $_POST["allow_incomplete_ingredients"] === 'true' ? '1' : '0';
     
     $response = [];
     foreach ($_POST as $key => $value) {