+++ core/list_ingredients_simple.php
@@ -48,10 +48,25 @@ while($res = mysqli_fetch_array($q)){
 }
 $i = 0;
 foreach ($ingredients as $ingredient) { 
-
-	$supp = getIngSupplier($ingredient['id'],1,$conn);
-	if($supp['price']){
-		
+	if(!$settings['allow_incomplete_ingredients']) {
+		$supp = getIngSupplier($ingredient['id'],1,$conn);
+		if($supp['price']){
+			$r['id'] = (int)$ingredient['id'];
+			$r['name'] = (string)$ingredient['name'];
+			$r['IUPAC'] = (string)$ingredient['INCI']?: '-';
+			$r['cas'] = (string)$ingredient['cas']?: '-';
+			$r['einecs'] = (string)$ingredient['einecs']?: '-';
+			$r['type'] = (string)$ingredient['type'] ?: 'Unknown';
+			$r['description'] = (string)$ingredient['odor'] ?: '-';
+			$r['physical_state'] = (int)$ingredient['physical_state'] ?: 1;
+			$r['profile'] = (string)$ingredient['profile'] ?: 'Unknown';
+			$r['stock'] = (float)number_format($supp['stock'], $settings['qStep']) ?: 0;
+			$r['mUnit'] = (string)$supp['mUnit'];
+			
+			$rx[]=$r;
+			$i++;
+		}
+	} else {
 		$r['id'] = (int)$ingredient['id'];
 		$r['name'] = (string)$ingredient['name'];
 		$r['IUPAC'] = (string)$ingredient['INCI']?: '-';
@@ -60,9 +75,9 @@ foreach ($ingredients as $ingredient) {
 		$r['type'] = (string)$ingredient['type'] ?: 'Unknown';
 		$r['description'] = (string)$ingredient['odor'] ?: '-';
 		$r['physical_state'] = (int)$ingredient['physical_state'] ?: 1;
-		$r['profile'] = (string)$ingredient['profile'] ?: 'Uknwown';
-		$r['stock'] = (float)number_format($supp['stock'], $settings['qStep']) ?: 0;
-		$r['mUnit'] = (string)$supp['mUnit'];
+		$r['profile'] = (string)$ingredient['profile'] ?: 'Unknown';
+		$r['stock'] = 0; // Default stock value
+		$r['mUnit'] = (string)$settings['mUnit'] ?? 'ml'; // Default measurement unit
 		
 		$rx[]=$r;
 		$i++;